Refactor create account to render separate component based on current step in account creation process
